Switch to gorilla sessions and flash (#171)
* Rename

* [SECURITY] Switch to html/template for escaping

* Provide generic 500 pages for when things go bad

* Use uints for apps and realms

* Add controller handlers for common responses

* Use csrf middleware and injector

* Switch to gorilla session flashes

* Use new APIs, ensure currentUser and currentRealm are set

* Upate templates to use new vars

* Update servers

* Add scripts in 500 to dismiss modals

* Spelling

* Fallback to Content-Type if Accept isn't provided

* Fix copy-pasta

* Move to middleware

* csrf -> csrfapi

* More renames

+++ pkg/database/realm.go
@@ -179,7 +179,7 @@ func (db *Database) GetRealmByName(name string) (*Realm, error) {
 	return &realm, nil
 }
 
-func (db *Database) GetRealm(realmID int64) (*Realm, error) {
+func (db *Database) GetRealm(realmID uint) (*Realm, error) {
 	var realm Realm
 	if err := db.db.Where("id = ?", realmID).First(&realm).Error; err != nil {
 		return nil, err
